requirements.txt
pillow==10.4.0
PyExifTool==0.5.6
